pyproject.toml
[tool.poetry]
name = "itosbot"
version = "0.1.0"
description = ""
authors = ["Alex <42418433+aiexz@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = { git = "https://github.com/aiogram/aiogram.git" }
pydantic-settings = "^2.0.2"
pillow = "^10.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
